Add files via upload
Sergio Moreno
Exploratory Analysis on how to build an Index ( Price vs Volume), 
How to build a tool  which can flag when to Buy or when to Sell a stock option based on short term vs. long term trend lines upwards vs. downwards crossover momentum. 
And a notebook to study a mutually exclusive output results on a montecarlo random run.